added attempt at a 3d discovery algorithm
Algorithm iterates through the transceivers on each agent to try and find where they line up. The agents are aligned with with transceiver 11 on each facing each other. A is lying flat on a table while B is on a platform about 6 inches higher. The platform can rotate and tilt. Transceiver 0 on B is currently facing directly up.
There is an issue with the algorithm where B might read "Hello" from a transceiver that B's selected transceiver is not facing. B then sends "H-ACK" but because B's selected transceiver is not facing A's selected transceiver, A does not read "H-ACK" and iterates while B listens for "ACK".